Added writing of track offsets

+++ tools/scp.h
@@ -34,8 +34,8 @@ struct scp_header
 
 extern void scp_writeheader(FILE *scpfile, const uint8_t rotations, const uint8_t starttrack, const uint8_t endtrack, const float rpm, const uint8_t sides);
 
-extern void scp_writetrack(FILE *scpfile, const int track, const unsigned char *rawtrackdata, const unsigned long rawdatalength, const unsigned int rotations);
+extern void scp_writetrack(FILE *scpfile, const uint8_t track, const unsigned char *rawtrackdata, const unsigned long rawdatalength, const uint8_t rotations);
 
-extern void scp_finalise(FILE *scpfile, const unsigned int endtrack);
+extern void scp_finalise(FILE *scpfile, const uint8_t endtrack);
 
 #endif
